improve and rearrange webhook processing
